Set 'fork id' to 0 as per current implementations
The 'fork id' could be changed if needed, but a value of 0 also works since
for transactions which set SIGHASH_FORKID, the adapted digest algorithm
is used, yielding a different signature.

A note is also added to the description of the new digest algorithm to
clarify that the GetForkID() function can be omitted in the UAHF, as it would
merely return 0 .

Specifying the GetForkID() function is still considered useful since the
document can remain applicable to chains which may want to set a non-zero
value as fork id.