Make protected transactions mandatory on UAHF chain (and enforce SCRIPT_VERIFY_STRICTENC) (#17)
* Make protected transactions mandatory on UAHF chain (and enforce SCRIPT_VERIFY_STRICTENC)

This enforces the use of SIGHASH_FORKID and SCRIPT_VERIFY_STRICTENC on
by UAHF-compatible clients once the fork has activated.

It makes the OP_RETURN protection obsolete, but it is not removed in order
to avoid a hard fork change w.r.t. existing deployments.

The strengthening of replay protections is a feature requested by miners
and exchanges to protect Bitcoin users from accidental loss and theft of
their Bitcoin Cash.

* Reword REQ-6-2 (as per deadalnix review comment)

Reworded positively and in bullet style similar to original

* Reword REQ-6-4

Thanks to Tom Zander for pointing out that its confusing.
Unfortunately, the specification situation around SCRIPT_VERIFY_STRICTENC
is also not such that we can point at a clear specification here.

Implementors should consider the code as reference for the moment.