Nav revert
Changing the nav to match the current site layout so there is not a difference between them in the intermin when both sites are live. Its a weird UX if the nav changes from page to page. I've stashed the changes I had made and can add back in when it makes sense