Updates to nav bar (#546)
* Updates to nav bar

Replaced white logo with color versoin and made clickable
Style edits to price ticker
Rearrangment of nav items

* formatting

* nav links rearrangment

removed community links to reddit pages replaced with developer links

* remove redundant code

* Remove Logo link from this PR

* Bring getting started back to own tab

* format fix

* Revert ticker.js

* revert ticker styling

* Revert color logo

Removing color logo from this PR so now its strictly nav items

* Remove uneccessary <></>on lang selector

* reorder array declerations to be in order

* Nav revert

Changing the nav to match the current site layout so there is not a difference between them in the intermin when both sites are live. Its a weird UX if the nav changes from page to page. I've stashed the changes I had made and can add back in when it makes sense

* Adding whitepaper back and explorer

* Adding explorer to current nav bar